src/cbers_cc/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
drf-flex-fields==1.0.2
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.2
